requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0
matplotlib>=3.6.0
seaborn>=0.12.0
joblib>=1.2.0
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0
geopandas>=0.13.0
pyproj>=3.5.0
jupyter>=1.0.0
notebook>=6.5.0
pyyaml>=6.0
requests>=2.28.0
shap>=0.41.0
lime>=0.2.0
imbalanced-learn>=0.10.0
pytest>=7.4.0
pytest-cov>=4.1.0